refactor: extract Injector to its own file
