+++ src/index.ts
@@ -3,7 +3,8 @@ export { createContainer } from "./container";
 export type { ContainerOptions } from "./containerOptions";
 export { AutoRegister, Inject, Injectable, InjectAll, Scoped } from "./decorators";
 export type { ClassDecorator, ClassFieldDecorator } from "./decorators/decorators";
-export { inject, injectAll, injectBy, Injector } from "./inject";
+export { inject, injectAll, injectBy } from "./inject";
+export { Injector } from "./injector";
 export type { Middleware, MiddlewareComposer } from "./middleware";
 export { applyMiddleware } from "./middleware";
 export type {
